CONC-365: Simple failover capabilities
It's now possible to specify multiple host
in connection parameters. The syntax is:

host="host1:port, host2:port, host3"

If no port was specified, the default port
will be used. IPv6 addresses have to specified
within square brackets.

mysql_real_connect will loop through the lists
of host until a connection was establised or
end of the list was reached.

The hostname may be specified as
- parameter in mysql_real_connect() api func
- in my.cnf (host=)
- in mariadb_dsn_connect
- via mysql_optionsv api func